create DB, connect to DB, add classes
The ChromaShare database was created on MongoDB.
The source code now connects to this database.
The classes.py file includes classes to represent user accounts
and photo posts, with the associated data for each type of object.
Using mongoengine, the classes are mongoengine documents. Their data
is mongoengine fields. They also have meta data to connect to the
database.